+++ src/server_discovery.rs
@@ -1,7 +1,7 @@
 use std::{
     collections::HashMap,
     io,
-    net::{Ipv4Addr, SocketAddr, ToSocketAddrs},
+    net::{Ipv4Addr, SocketAddr},
     sync::Arc,
     time::Duration,
 };
@@ -18,32 +18,15 @@ use tokio::{
 use tokio_stream::wrappers::IntervalStream;
 
 use crate::{
-    client_manager::ClientManagerCommand,
-    constants,
+    constants, create_default_proto_providers,
     proto::{DiscoveryRequest, DiscoveryResponse, ProviderInfo, ProviderType},
-    provider::{LMStudio, LlamaServer, Ollama, Provider, VLLM},
+    provider::Provider,
+    server_manager::ServerManagerCommand,
+    ALL_PROTO_PROVIDER_TYPES,
 };
 
-const RANDOM_UDP_PORT: u16 = 0;
-const DEFAULT_CLIENT_BROADCAST_INTERVAL: Duration = Duration::from_secs(5);
 const DEFAULT_SERVER_LIVENESS_INTERVAL: Duration = Duration::from_secs(10);
 const DISCOVERY_BUFFER_SIZE: usize = 1024;
-
-pub const DEFAULT_ALLOWED_PROVIDERS: &[ProviderType] = &[
-    ProviderType::Ollama,
-    ProviderType::Vllm,
-    ProviderType::LmStudio,
-    ProviderType::LlamaServer,
-];
-
-fn create_default_providers() -> HashMap<ProviderType, Arc<dyn Provider>> {
-    let mut providers: HashMap<ProviderType, Arc<dyn Provider>> = HashMap::new();
-    providers.insert(ProviderType::Ollama, Arc::new(Ollama::default()));
-    providers.insert(ProviderType::Vllm, Arc::new(VLLM::default()));
-    providers.insert(ProviderType::LmStudio, Arc::new(LMStudio::default()));
-    providers.insert(ProviderType::LlamaServer, Arc::new(LlamaServer::default()));
-    providers
-}
 
 pub fn convert_providers_to_server_proxy_ports(
     providers: &HashMap<ProviderType, Arc<dyn Provider>>,
@@ -55,17 +38,6 @@ pub fn convert_providers_to_server_proxy_ports(
     proxy_ports
 }
 
-/// Trait for abstracting network operations in ClientDiscovery.
-/// This allows for mocking network behavior in tests.
-#[async_trait]
-pub trait ClientDiscoveryNetwork: Send + Sync {
-    /// Send a discovery request to the broadcast address.
-    async fn send(&self, request: &DiscoveryRequest) -> io::Result<usize>;
-
-    /// Receive a discovery response from the network.
-    async fn recv(&self) -> io::Result<(DiscoveryResponse, SocketAddr)>;
-}
-
 /// Trait for abstracting network operations in ServerDiscovery.
 /// This allows for mocking network behavior in tests.
 #[async_trait]
@@ -75,66 +47,6 @@ pub trait ServerDiscoveryNetwork: Send + Sync {
 
     /// Send a discovery response to a specific address.
     async fn send(&self, response: &DiscoveryResponse, addr: SocketAddr) -> io::Result<usize>;
-}
-
-/// UDP-based implementation of ClientDiscoveryNetwork.
-pub struct UdpClientDiscoveryNetwork {
-    socket: UdpSocket,
-    server_port: u16,
-}
-
-impl UdpClientDiscoveryNetwork {
-    pub async fn new(server_port: u16) -> io::Result<Self> {
-        let socket = UdpSocket::bind((Ipv4Addr::UNSPECIFIED, RANDOM_UDP_PORT)).await?;
-        socket.set_broadcast(true)?;
-        Ok(Self {
-            socket,
-            server_port,
-        })
-    }
-
-    pub fn local_addr(&self) -> io::Result<SocketAddr> {
-        self.socket.local_addr()
-    }
-}
-
-#[async_trait]
-impl ClientDiscoveryNetwork for UdpClientDiscoveryNetwork {
-    async fn send(&self, request: &DiscoveryRequest) -> io::Result<usize> {
-        let mut buf = Vec::with_capacity(DISCOVERY_BUFFER_SIZE);
-        request.encode(&mut buf).map_err(|e| {
-            error!("Failed to encode DiscoveryRequest: {}", e);
-            io::Error::new(io::ErrorKind::InvalidData, e)
-        })?;
-
-        let len = self
-            .socket
-            .send_to(&buf, (Ipv4Addr::BROADCAST, self.server_port))
-            .await
-            .inspect_err(|error| error!("Client discovery error while sending: {}", error))?;
-
-        debug!("Client discovery sent {} bytes", len);
-
-        Ok(len)
-    }
-
-    async fn recv(&self) -> io::Result<(DiscoveryResponse, SocketAddr)> {
-        let mut buf = vec![0u8; DISCOVERY_BUFFER_SIZE];
-        let (len, addr) = self
-            .socket
-            .recv_from(&mut buf)
-            .await
-            .inspect_err(|error| error!("Client discovery error while receiving: {}", error))?;
-
-        debug!("Client discovery received {} bytes from {}", len, addr);
-
-        let response = DiscoveryResponse::decode(&buf[..len]).map_err(|e| {
-            debug!("Client discovery failed to decode message: {}", e);
-            io::Error::new(io::ErrorKind::InvalidData, e)
-        })?;
-
-        Ok((response, addr))
-    }
 }
 
 /// UDP-based implementation of ServerDiscoveryNetwork.
@@ -191,27 +103,12 @@ impl ServerDiscoveryNetwork for UdpServerDiscoveryNetwork {
     }
 }
 
-/// Trait for client-side discovery operations.
-/// This allows for different implementations of client discovery behavior.
-#[async_trait]
-pub trait ClientDiscovery: Send + Sync {
-    /// Run the client discovery process.
-    async fn run(&self, cmd_tx: &Sender<ClientManagerCommand>) -> anyhow::Result<()>;
-}
-
 /// Trait for server-side discovery operations.
 /// This allows for different implementations of server discovery behavior.
 #[async_trait]
 pub trait ServerDiscovery: Send + Sync {
     /// Run the server discovery process.
-    async fn run(&self) -> anyhow::Result<()>;
-}
-
-/// UDP-based implementation of ClientDiscovery.
-pub struct UdpClientDiscovery {
-    network: Arc<dyn ClientDiscoveryNetwork>,
-    broadcast_interval: std::time::Duration,
-    allowed_providers: Vec<ProviderType>,
+    async fn run(&self, cmd_tx: &Sender<ServerManagerCommand>) -> anyhow::Result<()>;
 }
 
 /// UDP-based implementation of ServerDiscovery.
@@ -222,119 +119,6 @@ pub struct UdpServerDiscovery {
     proxy_ports: HashMap<ProviderType, u16>,
     allowed_providers: Vec<ProviderType>,
     liveness_interval: std::time::Duration,
-}
-
-impl UdpClientDiscovery {
-    pub async fn new(
-        server_port: u16,
-        broadcast_interval: Duration,
-        allowed_providers: Vec<ProviderType>,
-    ) -> io::Result<Self> {
-        let network = UdpClientDiscoveryNetwork::new(server_port).await?;
-
-        Ok(Self {
-            network: Arc::new(network),
-            broadcast_interval,
-            allowed_providers,
-        })
-    }
-
-    pub async fn with_defaults() -> io::Result<Self> {
-        Self::new(
-            constants::OLLANA_SERVER_DEFAULT_DISCOVERY_PORT,
-            DEFAULT_CLIENT_BROADCAST_INTERVAL,
-            DEFAULT_ALLOWED_PROVIDERS.to_vec(),
-        )
-        .await
-    }
-
-    pub async fn with_allowed_providers(allowed_providers: Vec<ProviderType>) -> io::Result<Self> {
-        Self::new(
-            constants::OLLANA_SERVER_DEFAULT_DISCOVERY_PORT,
-            DEFAULT_CLIENT_BROADCAST_INTERVAL,
-            allowed_providers,
-        )
-        .await
-    }
-
-    pub fn with_network(
-        network: Arc<dyn ClientDiscoveryNetwork>,
-        broadcast_interval: Duration,
-        allowed_providers: Vec<ProviderType>,
-    ) -> Self {
-        Self {
-            network,
-            broadcast_interval,
-            allowed_providers,
-        }
-    }
-
-    /// Periodically broadcast discovery requests to find servers.
-    async fn broadcast_periodically(&self) -> anyhow::Result<()> {
-        let mut stream = IntervalStream::new(time::interval(self.broadcast_interval));
-
-        while stream.next().await.is_some() {
-            let request = DiscoveryRequest {
-                allowed_providers: self.allowed_providers.iter().map(|p| *p as i32).collect(),
-            };
-            let _ = self.network.send(&request).await;
-        }
-
-        Ok(())
-    }
-
-    /// Handle incoming discovery responses from servers.
-    async fn handle_messages(&self, cmd_tx: &Sender<ClientManagerCommand>) -> anyhow::Result<()> {
-        loop {
-            if let Ok((response, server_socket_addr)) = self.network.recv().await {
-                debug!(
-                    "Client discovery found a server with {} provider(s) at {}",
-                    response.provider_info.len(),
-                    server_socket_addr
-                );
-
-                for provider_info in response.provider_info {
-                    let provider_type = ProviderType::try_from(provider_info.provider_type)
-                        .map_err(|_| {
-                            anyhow::Error::msg(format!(
-                                "Unknown provider type: {}",
-                                provider_info.provider_type
-                            ))
-                        })?;
-                    let provider_port = provider_info.port as u16;
-                    let provider_socket_addr = (server_socket_addr.ip(), provider_port)
-                        .to_socket_addrs()?
-                        .next()
-                        .ok_or_else(|| {
-                            anyhow::Error::msg(format!(
-                                "Invalid server proxy address for provider {:?} on port {}",
-                                provider_type, provider_port
-                            ))
-                        })?;
-
-                    cmd_tx
-                        .send(ClientManagerCommand::Add(
-                            provider_type,
-                            provider_socket_addr,
-                        ))
-                        .await
-                        .unwrap_or(());
-                }
-            }
-        }
-    }
-}
-
-#[async_trait]
-impl ClientDiscovery for UdpClientDiscovery {
-    async fn run(&self, cmd_tx: &Sender<ClientManagerCommand>) -> anyhow::Result<()> {
-        info!("Running client discovery...");
-
-        tokio::select! {
-            val = self.broadcast_periodically() => val,
-            val = self.handle_messages(cmd_tx) => val,
-        }
-    }
 }
 
 impl UdpServerDiscovery {
@@ -358,13 +142,13 @@ impl UdpServerDiscovery {
     }
 
     pub async fn with_defaults() -> io::Result<Self> {
-        let providers = create_default_providers();
+        let providers = create_default_proto_providers();
         let proxy_ports = convert_providers_to_server_proxy_ports(&providers);
 
         Self::new(
             constants::OLLANA_SERVER_DEFAULT_DISCOVERY_PORT,
             providers,
-            DEFAULT_ALLOWED_PROVIDERS.to_vec(),
+            ALL_PROTO_PROVIDER_TYPES.to_vec(),
             proxy_ports,
             DEFAULT_SERVER_LIVENESS_INTERVAL,
         )
@@ -469,8 +253,8 @@ impl UdpServerDiscovery {
         }
     }
 
-    /// Periodically check the liveness of providers.
-    async fn run_liveness_check(&self) {
+    /// Periodically check the liveness of providers and send commands to ServerManager.
+    async fn run_liveness_check(&self, cmd_tx: &Sender<ServerManagerCommand>) {
         let mut stream = IntervalStream::new(time::interval(self.liveness_interval));
 
         while stream.next().await.is_some() {
@@ -490,6 +274,20 @@ impl UdpServerDiscovery {
                                     provider_type, provider_port
                                 );
                                 alive_providers.insert(*provider_type, provider_port);
+
+                                // Notify ServerManager to start proxy for this provider
+                                cmd_tx
+                                    .send(ServerManagerCommand::Start(
+                                        *provider_type,
+                                        provider_port,
+                                    ))
+                                    .await
+                                    .unwrap_or_else(|e| {
+                                        error!(
+                                            "Failed to send Start command to ServerManager: {}",
+                                            e
+                                        );
+                                    });
                             }
                         }
                         Ok(false) | Err(_) => {
@@ -498,6 +296,17 @@ impl UdpServerDiscovery {
                                     "Detected {:?} on port {} is no longer running, will stop responding for this provider",
                                     provider_type, provider_port
                                 );
+
+                                // Notify ServerManager to stop proxy for this provider
+                                cmd_tx
+                                    .send(ServerManagerCommand::Stop(*provider_type))
+                                    .await
+                                    .unwrap_or_else(|e| {
+                                        error!(
+                                            "Failed to send Stop command to ServerManager: {}",
+                                            e
+                                        );
+                                    });
                             }
                         }
                     }
@@ -509,12 +318,12 @@ impl UdpServerDiscovery {
 
 #[async_trait]
 impl ServerDiscovery for UdpServerDiscovery {
-    async fn run(&self) -> anyhow::Result<()> {
+    async fn run(&self, cmd_tx: &Sender<ServerManagerCommand>) -> anyhow::Result<()> {
         info!("Running server discovery...");
 
         tokio::select! {
             val = self.handle_messages() => val,
-            val = self.run_liveness_check() => Ok(val),
+            val = self.run_liveness_check(cmd_tx) => Ok(val),
         }
     }
 }
@@ -599,249 +408,6 @@ mod tests {
         }
     }
 
-    /// Mock implementation of ClientDiscoveryNetwork for testing.
-    pub struct MockClientDiscoveryNetwork {
-        responses: Arc<Mutex<Vec<(DiscoveryResponse, SocketAddr)>>>,
-    }
-
-    impl MockClientDiscoveryNetwork {
-        pub fn new(responses: Vec<(DiscoveryResponse, SocketAddr)>) -> Self {
-            Self {
-                responses: Arc::new(Mutex::new(responses)),
-            }
-        }
-    }
-
-    #[async_trait]
-    impl ClientDiscoveryNetwork for MockClientDiscoveryNetwork {
-        async fn send(&self, _request: &DiscoveryRequest) -> io::Result<usize> {
-            Ok(0)
-        }
-
-        async fn recv(&self) -> io::Result<(DiscoveryResponse, SocketAddr)> {
-            let mut responses = self.responses.lock().await;
-            if let Some(response) = responses.pop() {
-                Ok(response)
-            } else {
-                // Return an error to break out of the loop when no more responses
-                Err(io::Error::new(
-                    io::ErrorKind::WouldBlock,
-                    "no more responses",
-                ))
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn test_client_run_single_provider() {
-        let server_addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(192, 168, 1, 100)), 5000);
-        let provider_port = 11434u32;
-
-        let response = DiscoveryResponse {
-            provider_info: vec![ProviderInfo {
-                provider_type: ProviderType::Ollama as i32,
-                port: provider_port,
-            }],
-        };
-
-        let mock_network = Arc::new(MockClientDiscoveryNetwork::new(vec![(
-            response,
-            server_addr,
-        )]));
-
-        let client = UdpClientDiscovery::with_network(
-            mock_network,
-            DEFAULT_CLIENT_BROADCAST_INTERVAL,
-            DEFAULT_ALLOWED_PROVIDERS.to_vec(),
-        );
-
-        let (cmd_tx, mut cmd_rx) = mpsc::channel::<ClientManagerCommand>(32);
-
-        // Run client discovery in a separate task with a timeout
-        let handle = tokio::spawn(async move {
-            let _ = client.run(&cmd_tx).await;
-        });
-
-        // Wait for the command to be received
-        let cmd = tokio::time::timeout(Duration::from_millis(100), cmd_rx.recv())
-            .await
-            .expect("timeout waiting for command")
-            .expect("channel closed");
-
-        match cmd {
-            ClientManagerCommand::Add(provider_type, addr) => {
-                assert_eq!(provider_type, ProviderType::Ollama);
-                assert_eq!(addr.ip(), server_addr.ip());
-                assert_eq!(addr.port(), provider_port as u16);
-            }
-            _ => panic!("expected ClientManagerCommand::Add"),
-        }
-
-        handle.abort();
-    }
-
-    #[tokio::test]
-    async fn test_client_run_multiple_providers() {
-        let server_addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(192, 168, 1, 100)), 5000);
-
-        let response = DiscoveryResponse {
-            provider_info: vec![
-                ProviderInfo {
-                    provider_type: ProviderType::Ollama as i32,
-                    port: 11434,
-                },
-                ProviderInfo {
-                    provider_type: ProviderType::Vllm as i32,
-                    port: 8000,
-                },
-            ],
-        };
-
-        let mock_network = Arc::new(MockClientDiscoveryNetwork::new(vec![(
-            response,
-            server_addr,
-        )]));
-
-        let client = UdpClientDiscovery::with_network(
-            mock_network,
-            DEFAULT_CLIENT_BROADCAST_INTERVAL,
-            DEFAULT_ALLOWED_PROVIDERS.to_vec(),
-        );
-
-        let (cmd_tx, mut cmd_rx) = mpsc::channel::<ClientManagerCommand>(32);
-
-        let handle = tokio::spawn(async move {
-            let _ = client.run(&cmd_tx).await;
-        });
-
-        // Collect both commands
-        let mut received_providers = Vec::new();
-        for _ in 0..2 {
-            let cmd = tokio::time::timeout(Duration::from_millis(100), cmd_rx.recv())
-                .await
-                .expect("timeout waiting for command")
-                .expect("channel closed");
-
-            match cmd {
-                ClientManagerCommand::Add(provider_type, addr) => {
-                    assert_eq!(addr.ip(), server_addr.ip());
-                    received_providers.push((provider_type, addr.port()));
-                }
-                _ => panic!("expected ClientManagerCommand::Add"),
-            }
-        }
-
-        received_providers.sort_by_key(|&(_, port)| port);
-        assert_eq!(
-            received_providers,
-            vec![(ProviderType::Vllm, 8000), (ProviderType::Ollama, 11434)]
-        );
-
-        handle.abort();
-    }
-
-    #[tokio::test]
-    async fn test_client_run_from_multiple_servers() {
-        let server1_addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(192, 168, 1, 100)), 5000);
-        let server2_addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(192, 168, 1, 200)), 5000);
-
-        let response1 = DiscoveryResponse {
-            provider_info: vec![ProviderInfo {
-                provider_type: ProviderType::Ollama as i32,
-                port: 11434,
-            }],
-        };
-
-        let response2 = DiscoveryResponse {
-            provider_info: vec![ProviderInfo {
-                provider_type: ProviderType::Vllm as i32,
-                port: 8000,
-            }],
-        };
-
-        // Note: responses are popped from the end, so order is reversed
-        let mock_network = Arc::new(MockClientDiscoveryNetwork::new(vec![
-            (response2, server2_addr),
-            (response1, server1_addr),
-        ]));
-
-        let client = UdpClientDiscovery::with_network(
-            mock_network,
-            DEFAULT_CLIENT_BROADCAST_INTERVAL,
-            DEFAULT_ALLOWED_PROVIDERS.to_vec(),
-        );
-
-        let (cmd_tx, mut cmd_rx) = mpsc::channel::<ClientManagerCommand>(32);
-
-        let handle = tokio::spawn(async move {
-            let _ = client.run(&cmd_tx).await;
-        });
-
-        // First command should be from server1
-        let cmd1 = tokio::time::timeout(Duration::from_millis(100), cmd_rx.recv())
-            .await
-            .expect("timeout waiting for command")
-            .expect("channel closed");
-
-        match cmd1 {
-            ClientManagerCommand::Add(provider_type, addr) => {
-                assert_eq!(provider_type, ProviderType::Ollama);
-                assert_eq!(addr.ip(), server1_addr.ip());
-                assert_eq!(addr.port(), 11434);
-            }
-            _ => panic!("expected ClientManagerCommand::Add"),
-        }
-
-        // Second command should be from server2
-        let cmd2 = tokio::time::timeout(Duration::from_millis(100), cmd_rx.recv())
-            .await
-            .expect("timeout waiting for command")
-            .expect("channel closed");
-
-        match cmd2 {
-            ClientManagerCommand::Add(provider_type, addr) => {
-                assert_eq!(provider_type, ProviderType::Vllm);
-                assert_eq!(addr.ip(), server2_addr.ip());
-                assert_eq!(addr.port(), 8000);
-            }
-            _ => panic!("expected ClientManagerCommand::Add"),
-        }
-
-        handle.abort();
-    }
-
-    #[tokio::test]
-    async fn test_client_run_empty_provider_list() {
-        let server_addr = SocketAddr::new(IpAddr::V4(Ipv4Addr::new(192, 168, 1, 100)), 5000);
-
-        let response = DiscoveryResponse {
-            provider_info: vec![],
-        };
-
-        let mock_network = Arc::new(MockClientDiscoveryNetwork::new(vec![(
-            response,
-            server_addr,
-        )]));
-
-        let client = UdpClientDiscovery::with_network(
-            mock_network,
-            DEFAULT_CLIENT_BROADCAST_INTERVAL,
-            DEFAULT_ALLOWED_PROVIDERS.to_vec(),
-        );
-
-        let (cmd_tx, mut cmd_rx) = mpsc::channel::<ClientManagerCommand>(32);
-
-        let handle = tokio::spawn(async move {
-            let _ = client.run(&cmd_tx).await;
-        });
-
-        // Should not receive any commands when provider_info is empty
-        let result = tokio::time::timeout(Duration::from_millis(50), cmd_rx.recv()).await;
-        assert!(result.is_err(), "should timeout with no commands received");
-
-        handle.abort();
-    }
-
     /// Mock implementation of ServerDiscoveryNetwork for testing.
     pub struct MockServerDiscoveryNetwork {
         requests: Arc<Mutex<Vec<(DiscoveryRequest, SocketAddr)>>>,
@@ -883,6 +449,12 @@ mod tests {
         }
     }
 
+    /// Helper to create a dummy command sender for tests
+    fn create_dummy_cmd_tx() -> Sender<ServerManagerCommand> {
+        let (tx, _rx) = mpsc::channel::<ServerManagerCommand>(32);
+        tx
+    }
+
     /// Helper to create a UdpServerDiscovery with pre-populated alive providers
     async fn create_server_discovery_with_alive_providers(
         mock_network: Arc<MockServerDiscoveryNetwork>,
@@ -893,7 +465,7 @@ mod tests {
         let server = UdpServerDiscovery::with_network(
             mock_network,
             HashMap::new(), // providers not needed for handle_messages tests
-            DEFAULT_ALLOWED_PROVIDERS.to_vec(),
+            ALL_PROTO_PROVIDER_TYPES.to_vec(),
             proxy_ports,
             DEFAULT_SERVER_LIVENESS_INTERVAL,
         );
@@ -924,7 +496,11 @@ mod tests {
                 .await;
 
         // Run server discovery with a timeout - it will exit when mock returns error
-        let _ = tokio::time::timeout(Duration::from_millis(50), server.run()).await;
+        let _ = tokio::time::timeout(
+            Duration::from_millis(50),
+            server.run(&create_dummy_cmd_tx()),
+        )
+        .await;
 
         let sent_responses = mock_network.get_sent_responses().await;
         assert_eq!(sent_responses.len(), 1);
@@ -960,7 +536,11 @@ mod tests {
             create_server_discovery_with_alive_providers(mock_network.clone(), alive_providers)
                 .await;
 
-        let _ = tokio::time::timeout(Duration::from_millis(50), server.run()).await;
+        let _ = tokio::time::timeout(
+            Duration::from_millis(50),
+            server.run(&create_dummy_cmd_tx()),
+        )
+        .await;
 
         let sent_responses = mock_network.get_sent_responses().await;
         assert_eq!(sent_responses.len(), 1);
@@ -997,7 +577,11 @@ mod tests {
             create_server_discovery_with_alive_providers(mock_network.clone(), alive_providers)
                 .await;
 
-        let _ = tokio::time::timeout(Duration::from_millis(50), server.run()).await;
+        let _ = tokio::time::timeout(
+            Duration::from_millis(50),
+            server.run(&create_dummy_cmd_tx()),
+        )
+        .await;
 
         let sent_responses = mock_network.get_sent_responses().await;
         assert_eq!(sent_responses.len(), 1);
@@ -1033,7 +617,11 @@ mod tests {
             create_server_discovery_with_alive_providers(mock_network.clone(), alive_providers)
                 .await;
 
-        let _ = tokio::time::timeout(Duration::from_millis(50), server.run()).await;
+        let _ = tokio::time::timeout(
+            Duration::from_millis(50),
+            server.run(&create_dummy_cmd_tx()),
+        )
+        .await;
 
         let sent_responses = mock_network.get_sent_responses().await;
         // Should not send any response since there's no matching provider
@@ -1060,7 +648,11 @@ mod tests {
             create_server_discovery_with_alive_providers(mock_network.clone(), alive_providers)
                 .await;
 
-        let _ = tokio::time::timeout(Duration::from_millis(50), server.run()).await;
+        let _ = tokio::time::timeout(
+            Duration::from_millis(50),
+            server.run(&create_dummy_cmd_tx()),
+        )
+        .await;
 
         let sent_responses = mock_network.get_sent_responses().await;
         assert_eq!(sent_responses.len(), 0);
@@ -1092,7 +684,11 @@ mod tests {
             create_server_discovery_with_alive_providers(mock_network.clone(), alive_providers)
                 .await;
 
-        let _ = tokio::time::timeout(Duration::from_millis(50), server.run()).await;
+        let _ = tokio::time::timeout(
+            Duration::from_millis(50),
+            server.run(&create_dummy_cmd_tx()),
+        )
+        .await;
 
         let sent_responses = mock_network.get_sent_responses().await;
         assert_eq!(sent_responses.len(), 2);
@@ -1148,7 +744,11 @@ mod tests {
         assert!(server.alive_providers.lock().await.is_empty());
 
         // Run server discovery with timeout (includes liveness check)
-        let _ = tokio::time::timeout(Duration::from_millis(50), server.run()).await;
+        let _ = tokio::time::timeout(
+            Duration::from_millis(50),
+            server.run(&create_dummy_cmd_tx()),
+        )
+        .await;
 
         // Provider should now be alive
         let alive = server.alive_providers.lock().await;
@@ -1175,7 +775,11 @@ mod tests {
             .insert(ProviderType::Ollama, 11434);
 
         // Run server discovery (includes liveness check)
-        let _ = tokio::time::timeout(Duration::from_millis(50), server.run()).await;
+        let _ = tokio::time::timeout(
+            Duration::from_millis(50),
+            server.run(&create_dummy_cmd_tx()),
+        )
+        .await;
 
         // Provider should be removed
         let alive = server.alive_providers.lock().await;
@@ -1201,7 +805,11 @@ mod tests {
             .insert(ProviderType::Ollama, 11434);
 
         // Run server discovery (includes liveness check)
-        let _ = tokio::time::timeout(Duration::from_millis(50), server.run()).await;
+        let _ = tokio::time::timeout(
+            Duration::from_millis(50),
+            server.run(&create_dummy_cmd_tx()),
+        )
+        .await;
 
         // Provider should be removed on error
         let alive = server.alive_providers.lock().await;
@@ -1234,7 +842,11 @@ mod tests {
         );
 
         // Run server discovery (includes liveness check)
-        let _ = tokio::time::timeout(Duration::from_millis(50), server.run()).await;
+        let _ = tokio::time::timeout(
+            Duration::from_millis(50),
+            server.run(&create_dummy_cmd_tx()),
+        )
+        .await;
 
         // Only healthy providers should be alive
         let alive = server.alive_providers.lock().await;
@@ -1261,7 +873,11 @@ mod tests {
             create_server_discovery_with_mock_providers(providers, vec![ProviderType::Ollama]);
 
         // Run server discovery (includes liveness check)
-        let _ = tokio::time::timeout(Duration::from_millis(50), server.run()).await;
+        let _ = tokio::time::timeout(
+            Duration::from_millis(50),
+            server.run(&create_dummy_cmd_tx()),
+        )
+        .await;
 
         // Only allowed provider should be checked and added
         let alive = server.alive_providers.lock().await;
@@ -1288,7 +904,11 @@ mod tests {
 
         // Run server discovery - after multiple liveness check iterations, provider should be healthy
         // (first iteration: false -> not added, subsequent iterations: true -> added)
-        let _ = tokio::time::timeout(Duration::from_millis(50), server.run()).await;
+        let _ = tokio::time::timeout(
+            Duration::from_millis(50),
+            server.run(&create_dummy_cmd_tx()),
+        )
+        .await;
 
         // Provider should be alive (added during one of the later iterations)
         let alive = server.alive_providers.lock().await;
@@ -1314,7 +934,11 @@ mod tests {
 
         // Run server discovery - provider will be added then removed
         // (first iteration: true -> added, subsequent iterations: false -> removed)
-        let _ = tokio::time::timeout(Duration::from_millis(50), server.run()).await;
+        let _ = tokio::time::timeout(
+            Duration::from_millis(50),
+            server.run(&create_dummy_cmd_tx()),
+        )
+        .await;
 
         // Provider should not be alive (removed during later iterations)
         let alive = server.alive_providers.lock().await;
